+++ bin/bayesian_network_compiler.rs
@@ -4,8 +4,9 @@ extern crate rsgm;
 
 use clap::Parser;
 use rsdd::builder::cache::all_app::AllTable;
+use rsdd::builder::canonicalize::*;
 use rsdd::builder::decision_nnf_builder::DecisionNNFBuilder;
-use rsdd::builder::sdd_builder::{self, CompressionCanonicalizer};
+use rsdd::builder::sdd_builder::{self};
 use rsdd::repr::bdd::BddPtr;
 use rsdd::repr::ddnnf::DDNNFPtr;
 use rsdd::repr::dtree::DTree;
